fix: fix profile outlet routes
