+++ src/components/user-form/user-form.module.css
@@ -1,0 +1,37 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 24px;
+}
+
+.button {
+  max-width: fit-content;
+}
+
+.link {
+  text-decoration: none;
+  color: #4C4CFF;
+}
+
+.link:hover {
+  text-decoration: underline;
+  text-underline-position: under;
+}
+
+.buttonsSection {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+}
+
+.caption {
+  opacity: 0.4;
+}
+
+/* Fix styles of default components: */
+
+p {
+  margin-block-end: 0;
+}
