+++ src/pages/profile/profile.module.css
@@ -35,41 +35,3 @@
   color: #fff;
   text-decoration: none;
 }
-
-.form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 24px;
-}
-
-.button {
-  max-width: fit-content;
-}
-
-.link {
-  text-decoration: none;
-  color: #4C4CFF;
-}
-
-.link:hover {
-  text-decoration: underline;
-  text-underline-position: under;
-}
-
-.buttonsSection {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-}
-
-.caption {
-  opacity: 0.4;
-}
-
-/* Fix styles of default components: */
-
-p {
-  margin-block-end: 0;
-}
